refactoring PromotionService task perform

+++ src/main/java/school/faang/user_service/repository/UserRepository.java
@@ -30,7 +30,8 @@ public interface UserRepository extends JpaRepository<User, Long> {
     @Query(nativeQuery = true, value = """
             SELECT u.*
             FROM users u
-            LEFT JOIN user_promotion up ON u.id = up.user_id AND up.number_of_views > 0
+            LEFT JOIN user_promotion up
+            ON u.id = up.user_id AND up.number_of_views > 0
             ORDER BY
                 up.coefficient DESC NULLS LAST,
                 up.creation_date ASC,
@@ -38,7 +39,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
             OFFSET :offset
             LIMIT :limit
             """)
-    List<User> findAllSortedByPromotedUsersPerPage(@Param("offset") int offset, @Param("limit") int limit);
+    List<User> findAllSortedByPromotedUsersPerPage(@Param("offset") long offset, @Param("limit") long limit);
 
     @Query("""
             SELECT COUNT(f) FROM User u
